pyproject.toml
[project]
name = "autoprogcomp"
version = "0.1.0"
description = "Actualizar automaticamente el spreadsheet de notas del taller de progcomp"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-api-python-client>=2.164.0",
    "google-api-python-client-stubs>=1.29.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pydantic>=2.10.6",
    "python-dotenv==1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.396",
    "ruff>=0.11.0",
]

[tool.pyright]
include = ["app"]
exclude = [
    "**/__pycache__",
    "**/node_modules",
]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
pythonPlatform = "Linux"
useLibraryCodeForTypes = false
typeCheckingMode = "strict"

[tool.ruff]
include = ["pyproject.toml", "app/**/*.py"]
line-length = 120

[tool.ruff.lint]
select = ['E', 'F', 'W', 'A', 'PLC', 'PLE', 'PLW', 'I', 'RUF', 'UP']
